+++ lib/flutter_clean_calendar.dart
@@ -1,12 +1,18 @@
 library flutter_clean_calendar;
 
 import 'package:flutter/material.dart';
-import 'package:date_utils/date_utils.dart';
 import './simple_gesture_detector.dart';
 import './calendar_tile.dart';
-import 'event_dto.dart';
+import './clean_calendar_event.dart';
+import './date_utils.dart';
+import 'package:intl/intl.dart';
+import 'package:intl/date_symbol_data_local.dart';
+
+// Export NeatCleanCalendarEvent for using it in the application
+export './clean_calendar_event.dart';
 
 typedef DayBuilder(BuildContext context, DateTime day);
+typedef EventListBuilder(BuildContext context, List<CleanCalendarEvent> events);
 
 class Range {
   final DateTime from;
@@ -14,35 +20,109 @@ class Range {
   Range(this.from, this.to);
 }
 
+/// Clean Calndar's main class [Calendar]
+///
+/// This calls is responisble for controlling the look of the calnedar display as well
+/// as the action taken, when changing the month or tapping a date. It's higly configurable
+/// with its numerous properties.
+///
+/// [onDateSelected] is of type [ValueChanged<DateTime>] and it containes the callback function
+///     extecuted when tapping a date
+/// [onMonthChanged] is of type [ValueChanged<DateTime>] and it containes the callback function
+///     extecuted when changing to another month
+/// [onExpandStateChanged] is of type [ValueChanged<bool>] and it contains a callback function
+///     executed when the view changes to expanded or to condensed
+/// [onRangeSelected] contains a callback function of type [ValueChanged], that gets called on changes
+///     of the range (switch to next or previous week or month)
+/// [onEventSelected] is of type [ValueChanged<NeatCleanCalendarEvent>] and it contains a callback function
+///     executed when an event of the event list is selected
+/// [isExpandable] is a [bool]. With this parameter you can control, if the view can expand from week view
+///     to month view. Default is [false].
+/// [dayBuilder] can contain a [Widget]. If this property is not null (!= null), this widget will get used to
+///     render the calenar tiles (so you can customize the view)
+/// [eventListBuilder] can optionally contain a [Widget] that gets used to render the event list
+/// [hideArrows] is a bool. When set to [true] the arrows to navigate to the next or previous week/month in the
+///     top bar well get suppressed. Default is [false].
+/// [hideTodayIcon] is a bool. When set to [true] the dispaly of the Today-Icon (button to navigate to today) in the
+///     top bar well get suppressed. Default is [false].
+/// [hideBottomBar] at the moment has no function. Default is [false].
+/// [events] are of type [Map<DateTime, List<CleanCalendarEvent>>]. This data structure containes the events to display
+/// [selctedColor] this is the color, applied to the circle on the selcted day
+/// [todayColor] this is the color of the date of today
+/// [todayButtonText] is a [String]. With this property you can set the caption of the today icon (button to navigate to today).
+///     If left empty, the calendar will use the string "Today".
+/// [eventColor] lets you optionally specify the color of the event (dot). If the [CleanCaendarEvents] property color is not set, the
+///     calendar will use this parameter.
+/// [eventDoneColor] with this property you can define the color of "done" events, that is events in the past.
+/// [initialDate] is of type [DateTime]. It can contain an optional start date. This is the day, that gets initially selected
+///     by the calendar. The default is to not set this parameter. Then the calendar uses [DateTime.now()]
+/// [isExpanded] is a bool. If is us set to [true], the calendar gets rendered in month view.
+/// [weekDays] contains a [List<String>] defining the names of the week days, so that it is possible to name them according
+///     to your current locale.
+/// [locale] is a [String]. This setting gets used to format dates according to the current locale.
+/// [startOnMonday] is a [bool]. This parameter allows the calendar to determine the first day of the week.
+/// [dayOfWeekStyle] is a [TextStyle] for styling the text of the weekday names in the top bar.
+/// [bottomBarTextStyle] is a [TextStyle], that sets the style of the text in the bottom bar.
+/// [bottomBarArrowColor] can set the [Color] of the arrow to expand/compress the calendar in the bottom bar.
+/// [bottomBarColor] sets the [Color] of the bottom bar
+/// [expandableDateFormat] defines the formatting of the date in the bottom bar
 class Calendar extends StatefulWidget {
-  final ValueChanged<DateTime> onDateSelected;
-  final ValueChanged<DateTime> onMonthChanged;
-  final ValueChanged onRangeSelected;
+  final ValueChanged<DateTime>? onDateSelected;
+  final ValueChanged<DateTime>? onMonthChanged;
+  final ValueChanged<bool>? onExpandStateChanged;
+  final ValueChanged? onRangeSelected;
+  final ValueChanged<CleanCalendarEvent>? onEventSelected;
   final bool isExpandable;
-  final DayBuilder dayBuilder;
-  final bool showArrows;
-  final bool showTodayIcon;
-  final Map<DateTime, List<EventDto>> events;
-  final Color selectedColor;
-  final Color eventColor;
-  final Color eventDoneColor;
-  final DateTime initialDate;
+  final DayBuilder? dayBuilder;
+  final EventListBuilder? eventListBuilder;
+  final bool hideArrows;
+  final bool hideTodayIcon;
+  final Map<DateTime, List<CleanCalendarEvent>>? events;
+  final Color? selectedColor;
+  final Color? todayColor;
+  final String todayButtonText;
+  final Color? eventColor;
+  final Color? eventDoneColor;
+  final DateTime? initialDate;
   final bool isExpanded;
+  final List<String> weekDays;
+  final String? locale;
+  final bool startOnMonday;
+  final bool hideBottomBar;
+  final TextStyle? dayOfWeekStyle;
+  final TextStyle? bottomBarTextStyle;
+  final Color? bottomBarArrowColor;
+  final Color? bottomBarColor;
+  final String? expandableDateFormat;
 
   Calendar({
     this.onMonthChanged,
     this.onDateSelected,
     this.onRangeSelected,
+    this.onExpandStateChanged,
+    this.onEventSelected,
+    this.hideBottomBar: false,
     this.isExpandable: false,
     this.events,
     this.dayBuilder,
-    this.showTodayIcon: true,
-    this.showArrows: true,
+    this.eventListBuilder,
+    this.hideTodayIcon: false,
+    this.hideArrows: false,
     this.selectedColor,
+    this.todayColor,
+    this.todayButtonText: 'Today',
     this.eventColor,
     this.eventDoneColor,
     this.initialDate,
     this.isExpanded = false,
+    this.weekDays = const ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'],
+    this.locale = 'en_US',
+    this.startOnMonday = false,
+    this.dayOfWeekStyle,
+    this.bottomBarTextStyle,
+    this.bottomBarArrowColor,
+    this.bottomBarColor,
+    this.expandableDateFormat = 'EEEE MMMM dd, yyyy',
   });
 
   @override
@@ -51,28 +131,32 @@ class Calendar extends StatefulWidget {
 
 class _CalendarState extends State<Calendar> {
   final calendarUtils = Utils();
-  List<DateTime> selectedMonthsDays;
-  Iterable<DateTime> selectedWeeksDays;
+  late List<DateTime> selectedMonthsDays;
+  late Iterable<DateTime> selectedWeekDays;
   DateTime _selectedDate = DateTime.now();
-  String currentMonth;
-  bool isExpanded = false;
-  String displayMonth;
+  String? currentMonth;
+  late bool isExpanded;
+  String displayMonth = '';
   DateTime get selectedDate => _selectedDate;
-
-  
+  List<CleanCalendarEvent>? _selectedEvents;
 
   void initState() {
     super.initState();
-    _selectedDate = widget?.initialDate ?? DateTime.now();
-    isExpanded = widget?.isExpanded ?? false;
-    selectedMonthsDays = Utils.daysInMonth(_selectedDate);
-    var firstDayOfCurrentWeek = Utils.firstDayOfWeek(_selectedDate);
-    var lastDayOfCurrentWeek = Utils.lastDayOfWeek(_selectedDate);
-    selectedWeeksDays =
-        Utils.daysInRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek)
-            .toList()
-            .sublist(0, 7);
-    displayMonth = Utils.formatMonth(_selectedDate);
+    isExpanded = widget.isExpanded;
+    _selectedDate = widget.initialDate ?? DateTime.now();
+    selectedMonthsDays = _daysInMonth(_selectedDate);
+    selectedWeekDays = Utils.daysInRange(
+            _firstDayOfWeek(_selectedDate), _lastDayOfWeek(_selectedDate))
+        .toList();
+    initializeDateFormatting(widget.locale, null).then((_) => setState(() {
+          var monthFormat =
+              DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+          displayMonth =
+              '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+        }));
+    _selectedEvents = widget.events?[DateTime(
+            _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+        [];
   }
 
   Widget get nameAndIconRow {
@@ -80,7 +164,7 @@ class _CalendarState extends State<Calendar> {
     var leftArrow;
     var rightArrow;
 
-    if (widget.showArrows) {
+    if (!widget.hideArrows) {
       leftArrow = IconButton(
         onPressed: isExpanded ? previousMonth : previousWeek,
         icon: Icon(Icons.chevron_left),
@@ -94,9 +178,9 @@ class _CalendarState extends State<Calendar> {
       rightArrow = Container();
     }
 
-    if (widget.showTodayIcon) {
+    if (!widget.hideTodayIcon) {
       todayIcon = InkWell(
-        child: Text('Today'),
+        child: Text(widget.todayButtonText),
         onTap: resetToToday,
       );
     } else {
@@ -137,6 +221,7 @@ class _CalendarState extends State<Calendar> {
         ),
         child: Column(children: <Widget>[
           GridView.count(
+            childAspectRatio: 1.5,
             primary: false,
             shrinkWrap: true,
             crossAxisCount: 7,
@@ -151,18 +236,24 @@ class _CalendarState extends State<Calendar> {
   List<Widget> calendarBuilder() {
     List<Widget> dayWidgets = [];
     List<DateTime> calendarDays =
-        isExpanded ? selectedMonthsDays : selectedWeeksDays;
-
-    Utils.weekdays.forEach(
+        isExpanded ? selectedMonthsDays : selectedWeekDays as List<DateTime>;
+    widget.weekDays.forEach(
       (day) {
         dayWidgets.add(
           CalendarTile(
             selectedColor: widget.selectedColor,
+            todayColor: widget.todayColor,
             eventColor: widget.eventColor,
             eventDoneColor: widget.eventDoneColor,
-            events: widget.events[day],
+            events: widget.events![day],
             isDayOfWeek: true,
             dayOfWeek: day,
+            dayOfWeekStyle: widget.dayOfWeekStyle ??
+                TextStyle(
+                  color: widget.selectedColor,
+                  fontWeight: FontWeight.w500,
+                  fontSize: 11,
+                ),
           ),
         );
       },
@@ -175,7 +266,6 @@ class _CalendarState extends State<Calendar> {
       (day) {
         if (day.hour > 0) {
           day = day.toLocal();
-
           day = day.subtract(new Duration(hours: day.hour));
         }
 
@@ -188,13 +278,15 @@ class _CalendarState extends State<Calendar> {
         }
 
         if (this.widget.dayBuilder != null) {
+          // Use the dayBuilder widget passed as parameter to render the date tile
           dayWidgets.add(
             CalendarTile(
               selectedColor: widget.selectedColor,
+              todayColor: widget.todayColor,
               eventColor: widget.eventColor,
               eventDoneColor: widget.eventDoneColor,
-              events: widget.events[day],
-              child: this.widget.dayBuilder(context, day),
+              events: widget.events![day],
+              child: widget.dayBuilder!(context, day),
               date: day,
               onDateSelected: () => handleSelectedDateAndUserCallback(day),
             ),
@@ -203,9 +295,10 @@ class _CalendarState extends State<Calendar> {
           dayWidgets.add(
             CalendarTile(
                 selectedColor: widget.selectedColor,
+                todayColor: widget.todayColor,
                 eventColor: widget.eventColor,
                 eventDoneColor: widget.eventDoneColor,
-                events: widget.events[day],
+                events: widget.events![day],
                 onDateSelected: () => handleSelectedDateAndUserCallback(day),
                 date: day,
                 dateStyles: configureDateStyle(monthStarted, monthEnded),
@@ -218,17 +311,17 @@ class _CalendarState extends State<Calendar> {
     return dayWidgets;
   }
 
-  TextStyle configureDateStyle(monthStarted, monthEnded) {
-    TextStyle dateStyles;
-    final TextStyle body1Style = Theme.of(context).textTheme.body1;
+  TextStyle? configureDateStyle(monthStarted, monthEnded) {
+    TextStyle? dateStyles;
+    final TextStyle? body1Style = Theme.of(context).textTheme.bodyText2;
 
     if (isExpanded) {
-      final TextStyle body1StyleDisabled = body1Style.copyWith(
+      final TextStyle body1StyleDisabled = body1Style!.copyWith(
           color: Color.fromARGB(
         100,
-        body1Style.color.red,
-        body1Style.color.green,
-        body1Style.color.blue,
+        body1Style.color!.red,
+        body1Style.color!.green,
+        body1Style.color!.blue,
       ));
 
       dateStyles =
@@ -244,7 +337,7 @@ class _CalendarState extends State<Calendar> {
       return GestureDetector(
         onTap: toggleExpanded,
         child: Container(
-          color: Color.fromRGBO(0, 0, 0, 0.07),
+          color: widget.bottomBarColor ?? Color.fromRGBO(200, 200, 200, 0.2),
           height: 40,
           margin: EdgeInsets.only(top: 8.0),
           padding: EdgeInsets.all(0),
@@ -252,14 +345,24 @@ class _CalendarState extends State<Calendar> {
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: <Widget>[
               SizedBox(width: 40.0),
-              Text(Utils.fullDayFormat(selectedDate)),
+              Text(
+                DateFormat(widget.expandableDateFormat, widget.locale)
+                    .format(_selectedDate),
+                style: widget.bottomBarTextStyle ?? TextStyle(fontSize: 13),
+              ),
               IconButton(
-                onPressed: () {},
-                iconSize: 20.0,
+                onPressed: toggleExpanded,
+                iconSize: 25.0,
                 padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 10.0),
                 icon: isExpanded
-                    ? Icon(Icons.arrow_drop_up)
-                    : Icon(Icons.arrow_drop_down),
+                    ? Icon(
+                        Icons.arrow_drop_up,
+                        color: widget.bottomBarArrowColor ?? Colors.black,
+                      )
+                    : Icon(
+                        Icons.arrow_drop_down,
+                        color: widget.bottomBarArrowColor ?? Colors.black,
+                      ),
               ),
             ],
           ),
@@ -267,6 +370,90 @@ class _CalendarState extends State<Calendar> {
       );
     } else {
       return Container();
+    }
+  }
+
+  Widget get eventList {
+    if (widget.eventListBuilder == null) {
+      return Expanded(
+        child: _selectedEvents != null && _selectedEvents!.isNotEmpty
+            ? ListView.builder(
+                padding: EdgeInsets.all(0.0),
+                itemBuilder: (BuildContext context, int index) {
+                  final CleanCalendarEvent event = _selectedEvents![index];
+                  final String start =
+                      DateFormat('HH:mm').format(event.startTime).toString();
+                  final String end =
+                      DateFormat('HH:mm').format(event.endTime).toString();
+                  return Container(
+                    height: 60.0,
+                    child: InkWell(
+                      onTap: () {
+                        if (widget.onEventSelected != null) {
+                          widget.onEventSelected!(event);
+                        }
+                      },
+                      child: Row(
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        children: <Widget>[
+                          Expanded(
+                            flex: 5,
+                            child: Padding(
+                              padding: const EdgeInsets.all(4.0),
+                              child: Container(
+                                color: event.color,
+                              ),
+                            ),
+                          ),
+                          Expanded(
+                            flex: 75,
+                            child: Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Text(event.summary,
+                                      style: Theme.of(context)
+                                          .textTheme
+                                          .subtitle2),
+                                  Text(event.description)
+                                ],
+                              ),
+                            ),
+                          ),
+                          Expanded(
+                            flex: 20,
+                            child: Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.end,
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Text(start,
+                                      style: Theme.of(context)
+                                          .textTheme
+                                          .bodyText1),
+                                  Text(end,
+                                      style: Theme.of(context)
+                                          .textTheme
+                                          .bodyText1),
+                                ],
+                              ),
+                            ),
+                          )
+                        ],
+                      ),
+                    ),
+                  );
+                },
+                itemCount: _selectedEvents!.length,
+              )
+            : Container(),
+      );
+    } else {
+      // eventLiostBuilder is not null
+      return widget.eventListBuilder!(context, _selectedEvents!);
     }
   }
 
@@ -283,23 +470,33 @@ class _CalendarState extends State<Calendar> {
             expanded: calendarGridView,
             isExpanded: isExpanded,
           ),
-          expansionButtonRow
+          expansionButtonRow,
+          eventList
         ],
       ),
     );
   }
 
+  /// The function [resetToToday] is called on tap on the Today button in the top
+  /// position of the screen. It re-caclulates the range of dates, so that the
+  /// month view or week view changes to a range containing the current day.
   void resetToToday() {
     _selectedDate = DateTime.now();
-    var firstDayOfCurrentWeek = Utils.firstDayOfWeek(_selectedDate);
-    var lastDayOfCurrentWeek = Utils.lastDayOfWeek(_selectedDate);
+    var firstDayOfCurrentWeek = _firstDayOfWeek(_selectedDate);
+    var lastDayOfCurrentWeek = _lastDayOfWeek(_selectedDate);
 
     setState(() {
-      selectedWeeksDays =
+      selectedWeekDays =
           Utils.daysInRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek)
               .toList();
-      selectedMonthsDays = Utils.daysInMonth(_selectedDate);
-      displayMonth = Utils.formatMonth(_selectedDate);
+      selectedMonthsDays = _daysInMonth(_selectedDate);
+      var monthFormat =
+          DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+      displayMonth =
+          '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+      _selectedEvents = widget.events?[DateTime(
+              _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+          [];
     });
 
     _launchDateSelectionCallback(_selectedDate);
@@ -311,13 +508,16 @@ class _CalendarState extends State<Calendar> {
       var firstDateOfNewMonth = Utils.firstDayOfMonth(_selectedDate);
       var lastDateOfNewMonth = Utils.lastDayOfMonth(_selectedDate);
       updateSelectedRange(firstDateOfNewMonth, lastDateOfNewMonth);
-      selectedMonthsDays = Utils.daysInMonth(_selectedDate);
-      displayMonth = Utils.formatMonth(_selectedDate);
+      selectedMonthsDays = _daysInMonth(_selectedDate);
+      var monthFormat =
+          DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+      displayMonth =
+          '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+      _selectedEvents = widget.events?[DateTime(
+              _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+          [];
     });
-
-    if (widget.onMonthChanged != null) {
-      widget.onMonthChanged(_selectedDate);
-    }
+    _launchDateSelectionCallback(_selectedDate);
   }
 
   void previousMonth() {
@@ -326,26 +526,34 @@ class _CalendarState extends State<Calendar> {
       var firstDateOfNewMonth = Utils.firstDayOfMonth(_selectedDate);
       var lastDateOfNewMonth = Utils.lastDayOfMonth(_selectedDate);
       updateSelectedRange(firstDateOfNewMonth, lastDateOfNewMonth);
-      selectedMonthsDays = Utils.daysInMonth(_selectedDate);
-      displayMonth = Utils.formatMonth(_selectedDate);
+      selectedMonthsDays = _daysInMonth(_selectedDate);
+      var monthFormat =
+          DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+      displayMonth =
+          '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+      _selectedEvents = widget.events?[DateTime(
+              _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+          [];
     });
-
-    if (widget.onMonthChanged != null) {
-      widget.onMonthChanged(_selectedDate);
-    }
+    _launchDateSelectionCallback(_selectedDate);
   }
 
   void nextWeek() {
     setState(() {
       _selectedDate = Utils.nextWeek(_selectedDate);
-      var firstDayOfCurrentWeek = Utils.firstDayOfWeek(_selectedDate);
-      var lastDayOfCurrentWeek = Utils.lastDayOfWeek(_selectedDate);
+      var firstDayOfCurrentWeek = _firstDayOfWeek(_selectedDate);
+      var lastDayOfCurrentWeek = _lastDayOfWeek(_selectedDate);
       updateSelectedRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek);
-      selectedWeeksDays =
+      selectedWeekDays =
           Utils.daysInRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek)
-              .toList()
-              .sublist(0, 7);
-      displayMonth = Utils.formatMonth(_selectedDate);
+              .toList();
+      var monthFormat =
+          DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+      displayMonth =
+          '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+      _selectedEvents = widget.events?[DateTime(
+              _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+          [];
     });
     _launchDateSelectionCallback(_selectedDate);
   }
@@ -353,14 +561,19 @@ class _CalendarState extends State<Calendar> {
   void previousWeek() {
     setState(() {
       _selectedDate = Utils.previousWeek(_selectedDate);
-      var firstDayOfCurrentWeek = Utils.firstDayOfWeek(_selectedDate);
-      var lastDayOfCurrentWeek = Utils.lastDayOfWeek(_selectedDate);
+      var firstDayOfCurrentWeek = _firstDayOfWeek(_selectedDate);
+      var lastDayOfCurrentWeek = _lastDayOfWeek(_selectedDate);
       updateSelectedRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek);
-      selectedWeeksDays =
+      selectedWeekDays =
           Utils.daysInRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek)
-              .toList()
-              .sublist(0, 7);
-      displayMonth = Utils.formatMonth(_selectedDate);
+              .toList();
+      var monthFormat =
+          DateFormat('MMMM yyyy', widget.locale).format(_selectedDate);
+      displayMonth =
+          '${monthFormat[0].toUpperCase()}${monthFormat.substring(1)}';
+      _selectedEvents = widget.events?[DateTime(
+              _selectedDate.year, _selectedDate.month, _selectedDate.day)] ??
+          [];
     });
     _launchDateSelectionCallback(_selectedDate);
   }
@@ -368,7 +581,7 @@ class _CalendarState extends State<Calendar> {
   void updateSelectedRange(DateTime start, DateTime end) {
     Range _rangeSelected = Range(start, end);
     if (widget.onRangeSelected != null) {
-      widget.onRangeSelected(_rangeSelected);
+      widget.onRangeSelected!(_rangeSelected);
     }
   }
 
@@ -399,32 +612,90 @@ class _CalendarState extends State<Calendar> {
   void toggleExpanded() {
     if (widget.isExpandable) {
       setState(() => isExpanded = !isExpanded);
+      if (widget.onExpandStateChanged != null)
+        widget.onExpandStateChanged!(isExpanded);
     }
   }
 
   void handleSelectedDateAndUserCallback(DateTime day) {
-    var firstDayOfCurrentWeek = Utils.firstDayOfWeek(day);
-    var lastDayOfCurrentWeek = Utils.lastDayOfWeek(day);
+    var firstDayOfCurrentWeek = _firstDayOfWeek(day);
+    var lastDayOfCurrentWeek = _lastDayOfWeek(day);
+    // Check if the selected day falls into the next month. If this is the case,
+    // then we need to additionaly check, if a day in next year was selected.
     if (_selectedDate.month > day.month) {
-      previousMonth();
+      // Day in next year selected? Switch to next month.
+      if (_selectedDate.year < day.year) {
+        nextMonth();
+      } else {
+        previousMonth();
+      }
     }
+    // Check if the selected day falls into the last month. If this is the case,
+    // then we need to additionaly check, if a day in last year was selected.
     if (_selectedDate.month < day.month) {
-      nextMonth();
+      // Day in next last selected? Switch to next month.
+      if (_selectedDate.year > day.year) {
+        previousMonth();
+      } else {
+        nextMonth();
+      }
     }
     setState(() {
       _selectedDate = day;
-      selectedWeeksDays =
+      selectedWeekDays =
           Utils.daysInRange(firstDayOfCurrentWeek, lastDayOfCurrentWeek)
               .toList();
-      selectedMonthsDays = Utils.daysInMonth(day);
+      selectedMonthsDays = _daysInMonth(day);
+      _selectedEvents = widget.events?[_selectedDate] ?? [];
     });
     _launchDateSelectionCallback(day);
   }
 
   void _launchDateSelectionCallback(DateTime day) {
     if (widget.onDateSelected != null) {
-      widget.onDateSelected(day);
+      widget.onDateSelected!(day);
     }
+    if (widget.onMonthChanged != null) {
+      widget.onMonthChanged!(day);
+    }
+  }
+
+  _firstDayOfWeek(DateTime date) {
+    var day = DateTime.utc(
+        _selectedDate.year, _selectedDate.month, _selectedDate.day, 12);
+    if (widget.startOnMonday == true) {
+      day = day.subtract(Duration(days: day.weekday - 1));
+    } else {
+      // if the selected day is a Sunday, then it is already the first day of week
+      day = day.weekday == 7 ? day : day.subtract(Duration(days: day.weekday));
+    }
+    return day;
+  }
+
+  _lastDayOfWeek(DateTime date) {
+    return _firstDayOfWeek(date).add(new Duration(days: 7));
+  }
+
+  /// The function [_daysInMonth] takes the parameter [month] (which is of type [DateTime])
+  /// and calculates then all the days to be displayed in month view based on it. It returns
+  /// all that days in a [List<DateTime].
+  List<DateTime> _daysInMonth(DateTime month) {
+    var first = Utils.firstDayOfMonth(month);
+    var daysBefore = first.weekday;
+    var firstToDisplay = first.subtract(new Duration(days: daysBefore - 1));
+    var last = Utils.lastDayOfMonth(month);
+
+    var daysAfter = 7 - last.weekday;
+
+    // If the last day is sunday (7) the entire week must be rendered
+    if (daysAfter == 0) {
+      daysAfter = 7;
+    }
+
+    // Adding an extra day necessary. Otherwise the week with days in next month
+    // would always end on Saturdays.
+    var lastToDisplay = last.add(new Duration(days: daysAfter + 1));
+    return Utils.daysInRange(firstToDisplay, lastToDisplay).toList();
   }
 }
 
@@ -433,22 +704,22 @@ class ExpansionCrossFade extends StatelessWidget {
   final Widget expanded;
   final bool isExpanded;
 
-  ExpansionCrossFade({this.collapsed, this.expanded, this.isExpanded});
+  ExpansionCrossFade(
+      {required this.collapsed,
+      required this.expanded,
+      required this.isExpanded});
 
   @override
   Widget build(BuildContext context) {
-    return Flexible(
-      flex: 1,
-      child: AnimatedCrossFade(
-        firstChild: collapsed,
-        secondChild: expanded,
-        firstCurve: const Interval(0.0, 1.0, curve: Curves.fastOutSlowIn),
-        secondCurve: const Interval(0.0, 1.0, curve: Curves.fastOutSlowIn),
-        sizeCurve: Curves.decelerate,
-        crossFadeState:
-            isExpanded ? CrossFadeState.showSecond : CrossFadeState.showFirst,
-        duration: const Duration(milliseconds: 300),
-      ),
+    return AnimatedCrossFade(
+      firstChild: collapsed,
+      secondChild: expanded,
+      firstCurve: const Interval(0.0, 1.0, curve: Curves.fastOutSlowIn),
+      secondCurve: const Interval(0.0, 1.0, curve: Curves.fastOutSlowIn),
+      sizeCurve: Curves.decelerate,
+      crossFadeState:
+          isExpanded ? CrossFadeState.showSecond : CrossFadeState.showFirst,
+      duration: const Duration(milliseconds: 300),
     );
   }
 }