+++ lib/simple_gesture_detector.dart
@@ -8,32 +8,30 @@ import 'package:flutter/material.dart';
 class SimpleGestureDetector extends StatefulWidget {
   final Widget child;
   final SimpleSwipeConfig swipeConfig;
-  final HitTestBehavior behavior;
-  final VoidCallback onSwipeUp;
-  final VoidCallback onSwipeDown;
-  final VoidCallback onSwipeLeft;
-  final VoidCallback onSwipeRight;
+  final HitTestBehavior? behavior;
+  final VoidCallback? onSwipeUp;
+  final VoidCallback? onSwipeDown;
+  final VoidCallback? onSwipeLeft;
+  final VoidCallback? onSwipeRight;
 
   const SimpleGestureDetector({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
     this.swipeConfig = const SimpleSwipeConfig(),
     this.behavior,
     this.onSwipeUp,
     this.onSwipeDown,
     this.onSwipeLeft,
     this.onSwipeRight,
-  })  : assert(child != null),
-        assert(swipeConfig != null),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   _SimpleGestureDetectorState createState() => _SimpleGestureDetectorState();
 }
 
 class _SimpleGestureDetectorState extends State<SimpleGestureDetector> {
-  Offset _initialSwipeOffset;
-  Offset _finalSwipeOffset;
+  Offset? _initialSwipeOffset;
+  late Offset _finalSwipeOffset;
 
   void _onVerticalDragStart(DragStartDetails details) {
     _initialSwipeOffset = details.globalPosition;
@@ -45,15 +43,15 @@ class _SimpleGestureDetectorState extends State<SimpleGestureDetector> {
     if (widget.swipeConfig.swipeDetectionMoment ==
         SwipeDetectionMoment.onUpdate) {
       if (_initialSwipeOffset != null) {
-        final offsetDifference = _initialSwipeOffset.dy - _finalSwipeOffset.dy;
+        final offsetDifference = _initialSwipeOffset!.dy - _finalSwipeOffset.dy;
 
         if (offsetDifference.abs() > widget.swipeConfig.verticalThreshold) {
           _initialSwipeOffset = null;
           final isSwipeUp = offsetDifference > 0;
           if (isSwipeUp) {
-            widget.onSwipeUp();
+            widget.onSwipeUp!();
           } else {
-            widget.onSwipeDown();
+            widget.onSwipeDown!();
           }
         }
       }
@@ -67,15 +65,15 @@ class _SimpleGestureDetectorState extends State<SimpleGestureDetector> {
     }
 
     if (_initialSwipeOffset != null) {
-      final offsetDifference = _initialSwipeOffset.dy - _finalSwipeOffset.dy;
+      final offsetDifference = _initialSwipeOffset!.dy - _finalSwipeOffset.dy;
 
       if (offsetDifference.abs() > widget.swipeConfig.verticalThreshold) {
         _initialSwipeOffset = null;
         final isSwipeUp = offsetDifference > 0;
         if (isSwipeUp) {
-          widget.onSwipeUp();
+          widget.onSwipeUp!();
         } else {
-          widget.onSwipeDown();
+          widget.onSwipeDown!();
         }
       }
     }
@@ -91,15 +89,15 @@ class _SimpleGestureDetectorState extends State<SimpleGestureDetector> {
     if (widget.swipeConfig.swipeDetectionMoment ==
         SwipeDetectionMoment.onUpdate) {
       if (_initialSwipeOffset != null) {
-        final offsetDifference = _initialSwipeOffset.dx - _finalSwipeOffset.dx;
+        final offsetDifference = _initialSwipeOffset!.dx - _finalSwipeOffset.dx;
 
         if (offsetDifference.abs() > widget.swipeConfig.horizontalThreshold) {
           _initialSwipeOffset = null;
           final isSwipeLeft = offsetDifference > 0;
           if (isSwipeLeft) {
-            widget.onSwipeLeft();
+            widget.onSwipeLeft!();
           } else {
-            widget.onSwipeRight();
+            widget.onSwipeRight!();
           }
         }
       }
@@ -113,15 +111,15 @@ class _SimpleGestureDetectorState extends State<SimpleGestureDetector> {
     }
 
     if (_initialSwipeOffset != null) {
-      final offsetDifference = _initialSwipeOffset.dx - _finalSwipeOffset.dx;
+      final offsetDifference = _initialSwipeOffset!.dx - _finalSwipeOffset.dx;
 
       if (offsetDifference.abs() > widget.swipeConfig.horizontalThreshold) {
         _initialSwipeOffset = null;
         final isSwipeLeft = offsetDifference > 0;
         if (isSwipeLeft) {
-          widget.onSwipeLeft();
+          widget.onSwipeLeft!();
         } else {
-          widget.onSwipeRight();
+          widget.onSwipeRight!();
         }
       }
     }
@@ -165,7 +163,5 @@ class SimpleSwipeConfig {
     this.verticalThreshold = 50.0,
     this.horizontalThreshold = 50.0,
     this.swipeDetectionMoment = SwipeDetectionMoment.onEnd,
-  })  : assert(verticalThreshold != null),
-        assert(horizontalThreshold != null),
-        assert(swipeDetectionMoment != null);
+  });
 }