+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.1.0]
+- Introduction of 'CleanCalendarEvent' class. This makes the former syntax of the 'Map' that stores the events incompatible.
+
+## [1.0.2]
+- Breaking Changes: Changed naming properties from show to hide, i.e. `hideArrows` instead of `showArrows`
+- Fixed dependencies issues with newer flutter versions
+- Added more styling options
+- Ability to start month on Monday or Sunday. Sunday by default, you need to replace weekDays if changing this value
+- Ability to replace day of weeks when starting calendar on Monday 
+- Hide or show bottom bar and its color
+- Ability to change bottom bar color, arrow color and text style
+
 ## [1.0.1]
 
 - Ability to start the calendar with a date other than today